- Bring in all XRenderMetal files and macOS platform layer
- Include CMake build system and Dart shader toolchain scaffolding
- Merge Metal renderer implementation progress from MacPort branch

Made-with: Cursor

+++ CrySystem/Timer.cpp
@@ -244,7 +244,7 @@ float CTimer::MeasureTime(LPCSTR szComment)
   if(szComment==(LPCSTR)-1)
     szComment=0;
 
-  int nRecursionLevel = (int)m_pSystem->GetIRenderer()->EF_Query(EFQ_RecurseLevel) - 1;
+  int nRecursionLevel = (intptr_t)m_pSystem->GetIRenderer()->EF_Query(EFQ_RecurseLevel) - 1;
 	if(nRecursionLevel>0)
 		return 0;
